GCP/02_Code/02_CloudRun/Streaming/requirements.txt
torch
transformers
soundfile
numpy
flask
google-cloud-storage
google-cloud-firestore
